+++ web/styles/globals.css
@@ -43,14 +43,27 @@
     --color-primary-800: 19, 35, 76;
     --color-primary-900: 13, 24, 51;
 
+    --color-error-10: 255 252 252;
+    --color-error-20: 255 247 247;
+    --color-error-30: 255, 219, 220;
+    --color-error-100: 244 170 170;
+    --color-error-200: 220 62 62;
+    --color-error-500: 140 51 58;
+
     --color-background-100: 255, 255, 255; /* primary bg */
     --color-background-90: 247, 247, 247; /* secondary bg */
     --color-background-80: 232, 232, 232; /* tertiary bg */
+
+    --color-background-primary: var(--color-primary-10);
+    --color-background-error: var(--color-error-20);
 
     --color-text-100: 23, 23, 23; /* primary text */
     --color-text-200: 58, 58, 58; /* secondary text */
     --color-text-300: 82, 82, 82; /* tertiary text */
     --color-text-400: 163, 163, 163; /* placeholder text */
+
+    --color-text-primary: var(--color-primary-100);
+    --color-text-error: var(--color-error-200);
 
     --color-scrollbar: 163, 163, 163; /* scrollbar thumb */
 
@@ -58,6 +71,9 @@
     --color-border-200: 229, 229, 229; /* subtle border- 2 */
     --color-border-300: 212, 212, 212; /* strong border- 1 */
     --color-border-400: 185, 185, 185; /* strong border- 2 */
+
+    --color-border-primary: var(--color-primary-40);
+    --color-border-error: var(--color-error-100);
 
     --color-shadow-2xs: 0px 0px 1px 0px rgba(23, 23, 23, 0.06), 0px 1px 2px 0px rgba(23, 23, 23, 0.06),
       0px 1px 2px 0px rgba(23, 23, 23, 0.14);
@@ -151,7 +167,7 @@
 
     /* toast theme */
     --color-toast-success-text: 62, 155, 79;
-    --color-toast-error-text: 220, 62, 66;
+    --color-toast-error-text: var(--color-error-200);
     --color-toast-warning-text: 255, 186, 24;
     --color-toast-info-text: 51, 88, 212;
     --color-toast-loading-text: 28, 32, 36;
@@ -159,13 +175,13 @@
     --color-toast-tertiary-text: 96, 100, 108;
 
     --color-toast-success-background: 253, 253, 254;
-    --color-toast-error-background: 255, 252, 252;
+    --color-toast-error-background: var(--color-error-10);
     --color-toast-warning-background: 254, 253, 251;
     --color-toast-info-background: 253, 253, 254;
     --color-toast-loading-background: 253, 253, 254;
 
     --color-toast-success-border: 218, 241, 219;
-    --color-toast-error-border: 255, 219, 220;
+    --color-toast-error-border: var(--color-error-30);
     --color-toast-warning-border: 255, 247, 194;
     --color-toast-info-border: 210, 222, 255;
     --color-toast-loading-border: 224, 225, 230;
@@ -192,6 +208,15 @@
     --color-background-100: 25, 25, 25; /* primary bg */
     --color-background-90: 32, 32, 32; /* secondary bg */
     --color-background-80: 44, 44, 44; /* tertiary bg */
+
+    --color-background-primary: var(--color-background-90);
+    --color-background-error: var(--color-background-90);
+
+    --color-text-primary: var(--color-primary-40);
+    --color-text-error: var(--color-error-100);
+
+    --color-border-primary: var(--color-primary-200);
+    --color-border-error: var(--color-error-500);
 
     --color-shadow-2xs: 0px 0px 1px 0px rgba(0, 0, 0, 0.15), 0px 1px 3px 0px rgba(0, 0, 0, 0.5);
     --color-shadow-xs: 0px 0px 2px 0px rgba(0, 0, 0, 0.2), 0px 2px 4px 0px rgba(0, 0, 0, 0.5);
